Reorganized callback tests
